+++ globalsearchv2/global_search_v2_integration_test.go
@@ -22,7 +22,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/IBM/go-sdk-core/v4/core"
+	"github.com/IBM/go-sdk-core/v5/core"
 	common "github.com/IBM/platform-services-go-sdk/common"
 	"github.com/IBM/platform-services-go-sdk/globalsearchv2"
 	"github.com/google/uuid"