+++ openservicebrokerv1/open_service_broker_v1_examples_test.go
@@ -159,7 +159,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nGetServiceInstanceState() result:\n%s\n", string(b))
 
 			// end-get_service_instance_state
 
@@ -182,7 +182,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nReplaceServiceInstanceState() result:\n%s\n", string(b))
 
 			// end-replace_service_instance_state
 
@@ -218,7 +218,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nReplaceServiceInstance() result:\n%s\n", string(b))
 
 			// end-replace_service_instance
 
@@ -256,7 +256,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nUpdateServiceInstance() result:\n%s\n", string(b))
 
 			// end-update_service_instance
 
@@ -275,7 +275,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nListCatalog() result:\n%s\n", string(b))
 
 			// end-list_catalog
 
@@ -299,7 +299,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nGetLastOperation() result:\n%s\n", string(b))
 
 			// end-get_last_operation
 
@@ -333,7 +333,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nReplaceServiceBinding() result:\n%s\n", string(b))
 
 			// end-replace_service_binding
 
@@ -356,7 +356,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 				panic(err)
 			}
 			b, _ := json.MarshalIndent(result, "", "  ")
-			fmt.Println(string(b))
+			fmt.Printf("\nDeleteServiceInstance() result:\n%s\n", string(b))
 
 			// end-delete_service_instance
 
@@ -379,6 +379,7 @@ var _ = Describe(`OpenServiceBrokerV1 Examples Tests`, func() {
 			if err != nil {
 				panic(err)
 			}
+			fmt.Printf("\nDeleteServiceBinding() response status code: %d\n", response.StatusCode)
 
 			// end-delete_service_binding